Adapting fulltext for use with common utilities
