+++ pdf_extraction/common/utilities.py
@@ -4,12 +4,29 @@
 #==============================================================================
 #Libraries
 #==============================================================================
-import requests #upload_task
+#for label studio interactions: upload_task
+import requests 
 
+#for fetching and saving docs
+import os 
+from metapub import PubMedFetcher 
+from metapub import FindIt 
+
+#PDF extraction: extract_text_from_pdf
+import fitz  # PyMuPDF
+
+#Text preprocessing: preprocess_text, identify_sections, 
+import re
+import nltk
+from nltk.tokenize import sent_tokenize
+# Download NLTK data files
+nltk.download('punkt')
+
+#for NLP/NER work: extract_entities, find_entity_groups
+import spacy 
 #==============================================================================
 # Functions for dealing with label studio 
 #==============================================================================
-
 # Function to upload text data to Label Studio
 def upload_task(text, project_id):
 	import_url = f'{LABEL_STUDIO_URL}/api/projects/{project_id}/import'
@@ -32,6 +49,84 @@ def upload_task(text, project_id):
 	except requests.exceptions.JSONDecodeError as e:
 		print("Failed to decode JSON:", e)
 		return None
+
+#==============================================================================
+# Functions for fetching and processing docs
+#==============================================================================
+# Function to fetch and save full text articles
+def get_full_text(articles, save_directory):
+    # Ensure the save directory exists
+    os.makedirs(save_directory, exist_ok=True)
+    # # Create a PubMedFetcher instance
+    # fetcher = PubMedFetcher()
+    total_attempted = 0
+    total_successful = 0
+    for article in articles:
+        total_attempted += 1
+        try:
+            # Get the PMID of the article
+            pmid = article.pmid
+            # Use FindIt to get the URL of the free open access article full text
+            url = FindIt(pmid).url
+            if url:
+                # Get the full text content
+                response = requests.get(url)
+                response.raise_for_status()  # Raise an error for bad status codes
+                # Create a filename for the article based on its PMID
+                filename = f"{pmid}.pdf"
+                file_path = os.path.join(save_directory, filename)
+                # Save the full text to the specified directory
+                with open(file_path, 'wb') as file:
+                    file.write(response.content)
+                print(f"Downloaded full text for PMID {pmid} to {file_path}")
+                total_successful += 1
+            else:
+                print(f"No free full text available for PMID {pmid}")
+        except Exception as e:
+            print(f"An error occurred for PMID {pmid}: {e}")
+    print(f"Total articles attempted: {total_attempted}")
+    print(f"Total articles successfully retrieved: {total_successful}")
+
+
+#Extract Text from PDF
+def extract_text_from_pdf(pdf_path):
+    doc = fitz.open(pdf_path)
+    text = ""
+    for page_num in range(len(doc)):
+        page = doc.load_page(page_num)
+        text += page.get_text()
+    return text
+
+#Preprocess Text
+def preprocess_text(text):
+    # Remove References/Bibliography and Acknowledgements sections
+    text = re.sub(r'\bREFERENCES\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
+    text = re.sub(r'\bACKNOWLEDGEMENTS\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
+    text = re.sub(r'\bBIBLIOGRAPHY\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
+    # Tokenize text into sentences
+    sentences = sent_tokenize(text)
+    return sentences
+
+def identify_sections(sentences, section_mapping):
+    sections = {'abstract','introduction','methods','results','discussion' }
+    # Initialize the sections dictionary with each section name as a key and an empty list as the value
+    sections = {section: [] for section in sections}
+    current_section = None
+
+    # Enhanced regex to match section headers
+    section_header_pattern = re.compile(r'\b(Abstract|Introduction|Methods|Materials and Methods|Results|Discussion|Conclusion|Background|Summary)\b', re.IGNORECASE)
+    for sentence in sentences:
+        # Check if the sentence is a section header
+        header_match = section_header_pattern.search(sentence)
+        if header_match:
+            section_name = header_match.group(1).lower()
+            normalized_section = section_mapping.get(section_name, section_name)
+            current_section = normalized_section
+            sections[current_section].append(sentence)
+            print(f"Matched Section Header: {header_match}")  # Debugging line
+        elif current_section:
+            sections[current_section].append(sentence)
+    return sections
 
 
 #==============================================================================
