+++ pdf_extraction/fulltext/extract_responses_txt_v2.py
@@ -53,6 +53,15 @@ import spacy
 #Post-NER processing
 from collections import defaultdict
 
+#The shared custom definitions
+#NOTE: This line might have to be modified as structure changes and 
+#we move towards deployment
+## Add the project root directory to sys.path
+#sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+import sys
+sys.path.append(os.path.abspath('./../'))  
+import common.utilities
+
 #==============================================================================
 #Set paths: 
 #==============================================================================
