+++ pdf_extraction/fulltext/meta_analyze_get_full.py
@@ -25,24 +25,23 @@ NCBI_API_KEY = "f2857b2abca3fe365c756aeb647e06417b08"
 import pandas as pd
 
 #For references 
-from metapub import PubMedFetcher 
-from metapub import FindIt 
 import os
 os.environ['NCBI_API_KEY'] = NCBI_API_KEY
 
 #For label studio
-import requests
 from label_studio_sdk import Client
-
-#PDF extraction
-import fitz  # PyMuPDF
-#Text preprocessing
-import re
-import nltk
-from nltk.tokenize import sent_tokenize
 
 #NLP
 import spacy
+
+#The shared custom definitions
+#NOTE: This line might have to be modified as structure changes and 
+#we move towards deployment
+## Add the project root directory to sys.path
+#sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+import sys
+sys.path.append(os.path.abspath('./../'))  
+import common.utilities
 #==============================================================================
 #Fetch records from PubMed
 #==============================================================================
@@ -66,47 +65,13 @@ for pmid in pmids:
     articles.append(article)
 
 #==============================================================================
-#Try to get full text PDFs 
+#Try to get full text PDFs from NCBI
 #==============================================================================
 # Directory to save the full text articles
-save_directory = './papers/'
-
-# Function to fetch and save full text articles
-def get_full_text(articles, save_directory):
-    # Ensure the save directory exists
-    os.makedirs(save_directory, exist_ok=True)
-    # # Create a PubMedFetcher instance
-    # fetcher = PubMedFetcher()
-    total_attempted = 0
-    total_successful = 0
-    for article in articles:
-        total_attempted += 1
-        try:
-            # Get the PMID of the article
-            pmid = article.pmid
-            # Use FindIt to get the URL of the free open access article full text
-            url = FindIt(pmid).url
-            if url:
-                # Get the full text content
-                response = requests.get(url)
-                response.raise_for_status()  # Raise an error for bad status codes
-                # Create a filename for the article based on its PMID
-                filename = f"{pmid}.pdf"
-                file_path = os.path.join(save_directory, filename)
-                # Save the full text to the specified directory
-                with open(file_path, 'wb') as file:
-                    file.write(response.content)
-                print(f"Downloaded full text for PMID {pmid} to {file_path}")
-                total_successful += 1
-            else:
-                print(f"No free full text available for PMID {pmid}")
-        except Exception as e:
-            print(f"An error occurred for PMID {pmid}: {e}")
-    print(f"Total articles attempted: {total_attempted}")
-    print(f"Total articles successfully retrieved: {total_successful}")
+save_directory = './../papers/'
 
 # Fetch and save full text articles
-get_full_text(articles, save_directory)
+common.utilities.get_full_text(articles, save_directory)
 
 #==============================================================================
 # Create project to fine-tune an NER to pull useful info from abstracts
@@ -159,30 +124,7 @@ print("Response Text:", response.text)
 # 4. Upload these to label studio
 #==============================================================================
 #Step 1: Extract Text from PDF
-
-def extract_text_from_pdf(pdf_path):
-    doc = fitz.open(pdf_path)
-    text = ""
-    for page_num in range(len(doc)):
-        page = doc.load_page(page_num)
-        text += page.get_text()
-    return text
-
 #Step 2: Preprocess Text
-# Download NLTK data files
-nltk.download('punkt')
-
-def preprocess_text(text):
-    # Remove References/Bibliography and Acknowledgements sections
-    text = re.sub(r'\bREFERENCES\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
-    text = re.sub(r'\bACKNOWLEDGEMENTS\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
-    text = re.sub(r'\bBIBLIOGRAPHY\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
-    # Tokenize text into sentences
-    sentences = sent_tokenize(text)
-    return sentences
-
-
-
 #Step 3: Identify Sections
 #Define a mapping for section variations
 section_mapping = {
@@ -199,51 +141,7 @@ section_mapping = {
     'conclusion': 'discussion',
     'summary': 'discussion'
 }
-
-def identify_sections(sentences):
-    sections = {'abstract','introduction','methods','results','discussion' }
-    # Initialize the sections dictionary with each section name as a key and an empty list as the value
-    sections = {section: [] for section in sections}
-    current_section = None
-
-    # Enhanced regex to match section headers
-    section_header_pattern = re.compile(r'\b(Abstract|Introduction|Methods|Materials and Methods|Results|Discussion|Conclusion|Background|Summary)\b', re.IGNORECASE)
-    for sentence in sentences:
-        # Check if the sentence is a section header
-        header_match = section_header_pattern.search(sentence)
-        if header_match:
-            section_name = header_match.group(1).lower()
-            normalized_section = section_mapping.get(section_name, section_name)
-            current_section = normalized_section
-            sections[current_section].append(sentence)
-            print(f"Matched Section Header: {header_match}")  # Debugging line
-        elif current_section:
-            sections[current_section].append(sentence)
-    return sections
-
-
 #Step 4: Upload text data to Label Studio
-def upload_task(text, project_id):
-    import_url = f'{LABEL_STUDIO_URL}/api/projects/{project_id}/import'
-    print("Import URL:", import_url)
-    response = requests.post(
-        import_url,
-        headers={'Authorization': f'Token {API_KEY}'},
-        json=[{
-            'data': {
-                'text': text
-            }
-        }]
-    )
-    print("Status Code:", response.status_code)
-    print("Response Text:", response.text)
-    try:
-        response_json = response.json()
-        print(response_json)
-        return response_json
-    except requests.exceptions.JSONDecodeError as e:
-        print("Failed to decode JSON:", e)
-        return None
 
 # Now put all of this together and do it! 
 # Keep track of PDFs that have already been processed
@@ -267,23 +165,23 @@ new_pdfs = current_pdfs - processed_pdfs
 # Process the new PDFs
 for pdf in new_pdfs:
     #1.
-    pdf_path = "./papers/" + pdf
-    pdf_text = extract_text_from_pdf(pdf_path)
+    pdf_path = "./../papers/" + pdf
+    pdf_text = common.utilities.extract_text_from_pdf(pdf_path)
     #2. 
-    sentences = preprocess_text(pdf_text)
+    sentences = common.utilities.preprocess_text(pdf_text)
     #3.
-    sections = identify_sections(sentences)
+    sections = common.utilities.identify_sections(sentences,section_mapping)
     #4.
     #Get the methods and upload
     methods_text = " ".join(sections.get('methods', [])) 
     if methods_text: # Check it exists
-        upload_task(methods_text, PROJECT_ID)
+        common.utilities.upload_task(methods_text, PROJECT_ID)
     else:
         print(f"No Methods found for article with PMID: {pdf}")
     #Get the reults and upload
     results_text = " ".join(sections.get('results', []))
     if results_text: # Check it exists
-        upload_task(results_text, PROJECT_ID)
+        common.utilities.upload_task(results_text, PROJECT_ID)
     else:
         print(f"No Results found for article with PMID: {pdf}")
     #Keep track of processed PDFs
